inferences/chatbot_v1beta/requirements.txt
python-dotenv
google-generativeai
streamlit
google-cloud-aiplatform